レイアウト変更

+++ app/src/main/java/systems/ohno/h_money/model/Speed.kt
@@ -2,7 +2,7 @@ package systems.ohno.h_money.model
 
 
 data class Speed (
-        val hour_speed: Int,
+        val hour_speed: Float,
         val remaing_amount: Int,
         val average_speed:Int
 )
